Move reusable hook to lib package
